updated commands.py

+++ app/commands.py
@@ -7,20 +7,68 @@ from app.profile.models import *
 
 db_cli = AppGroup("db")
 
-@click.command("flushDb")
+@db_cli.command("drop", help="Drop all tables")
+@click.option("--only", help="Only drop this table")
 @with_appcontext
-def flushDb():
-    """Clear all tables from the database
+def drop(only):
+    """Drop all tables from the database.
     
-    After this is complete, the db.create_all function needs to run to
-    create all the tables again
+    If the --only <table> option is used, only that table is dropped
+    """
+    meta = db.metadata
+    if only:
+        for table in reversed(meta.sorted_tables):
+            if str(table) == only:
+                print("Dropped table %s" % table)
+                table.drop(db.engine)
+                meta.remove(table)
+                return
+    else:
+        for table in reversed(meta.sorted_tables):
+            print("Dropped table %s" % table)
+            table.drop(db.engine)
+            meta.remove(table)
+        meta.drop_all(db.engine)
+
+@db_cli.command("clear", help="Clear all tables")
+@click.option("--only", help="Only clear this table")
+@with_appcontext
+def clear(only):
+    """Clear the contents of all tables
+    
+    If the --only <table> option is used, only that table is cleared
+    """
+    meta = db.metadata
+    if only:
+        for table in reversed(meta.sorted_tables):
+            if str(table) == only:
+                print("Cleared table %s" % table)
+                db.session.execute(table.delete())
+            db.session.commit()
+    else:
+        for table in reversed(meta.sorted_tables):
+            print("Cleared table %s" % table)
+            db.session.execute(table.delete())
+        db.session.commit()
+
+@db_cli.command("init")
+@with_appcontext
+def init():
+    """Initiate all tables
+
+    """
+    db.create_all()
+    print("Done")
+
+@db_cli.command("list")
+@with_appcontext
+def list():
+    """List all tables (BROKEN)
     """
     meta = db.metadata
     for table in reversed(meta.sorted_tables):
-        print("Clear table %s" % table)
-        db.session.execute(table.delete())
-    db.session.commit()
-    
+        print(table)
+
 @db_cli.command("stats", help="List number of rows for each table")
 @with_appcontext
 def stats():
